+++ packages/renderer/src/example-videos.ts
@@ -40,4 +40,5 @@ export const exampleVideos = {
 	mp4withmp3: path.join(examplePackage, 'public', 'mp4-mp3.mp4'),
 	av1bbb: path.join(examplePackage, 'public', 'av1-bbb.webm'),
 	av1mp4: path.join(examplePackage, 'public', 'av1-bbb.mp4'),
+	av1mp4WithColr: path.join(examplePackage, 'public', 'av1.mp4'),
 };
